pyproject.toml
[project]
name = "di"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "pillow>=12.1.1",
    "requests>=2.32.5",
]
